voting stand ranking controller, service, etc

+++ src/main/resources/schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS stand(
    short_description VARCHAR NOT NULL,
    description VARCHAR NOT NULL,
    cover VARCHAR NOT NULL,
-   ranking INTEGER
+   ranking FLOAT
 );
 
 CREATE TABLE IF NOT EXISTS stand_pictures(
